requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
flet==0.25.1
flet-desktop==0.25.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jwt==1.3.1
oauthlib==3.2.2
pycparser==2.22
repath==0.9.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
